unit test for btc tx serialization
